+++ frontend/packages/process-editor/src/components/ConfigPanel/ConfigContent/EditTaskId/EditTaskId.tsx
@@ -4,13 +4,11 @@ import { checkForInvalidCharacters } from '../../../../utils/configPanelUtils';
 import { StudioToggleableTextfield } from '@studio/components';
 import { KeyVerticalIcon } from '@studio/icons';
 import { useBpmnContext } from '../../../../contexts/BpmnContext';
-
-import type { MetaDataForm } from '../../../../contexts/BpmnConfigPanelContext';
 import { useBpmnConfigPanelFormContext } from '../../../../contexts/BpmnConfigPanelContext';
 import type Modeling from 'bpmn-js/lib/features/modeling/Modeling';
-
 import classes from './EditTaskId.module.css';
 import { useTaskIds } from '../../../../hooks/useTaskIds';
+import type { MetaDataForm } from 'app-shared/types/BpmnMetaDataForm';
 
 type EditTaskIdProps = HTMLAttributes<HTMLDivElement>;
 export const EditTaskId = ({ ...rest }: EditTaskIdProps): React.ReactElement => {
