+++ frontend/packages/shared/src/types/BpmnMetaDataForm.ts
@@ -1,0 +1,10 @@
+export type MetaDataForm = {
+  taskIdChanges?: Array<{
+    oldId: string;
+    newId: string;
+  }>;
+  dataTypeChangeDetails?: {
+    newDataType: string;
+    connectedTaskId: string;
+  };
+};
